move HighTable to its own component
