+++ src/components/HighTable/HighTable.tsx
@@ -1,27 +1,17 @@
 import React, { MouseEvent, useCallback, useEffect, useMemo, useRef, useState } from 'react'
-import Cell from './components/Cell/Cell.js'
-import Row from './components/Row/Row.js'
-import RowHeader from './components/RowHeader/RowHeader.js'
-import TableCorner from './components/TableCorner/TableCorner.js'
-import TableHeader, { cellStyle } from './components/TableHeader/TableHeader.js'
-import { DataFrame } from './helpers/dataframe.js'
-import { PartialRow } from './helpers/row.js'
-import { Selection, SortIndex, areAllSelected, computeNewSelection, isSelected, toggleAll } from './helpers/selection.js'
-import { OrderBy, areEqualOrderBy } from './helpers/sort.js'
-import { formatRowNumber, rowError } from './helpers/text.js'
-import { useInputState } from './hooks/useInputState.js'
-import { stringify as stringifyDefault } from './utils/stringify.js'
-import { throttle } from './utils/throttle.js'
-export { arrayDataFrame, sortableDataFrame } from './helpers/dataframe.js'
-export type { DataFrame } from './helpers/dataframe.js'
-export { asyncRows, awaitRow, awaitRows, resolvableRow } from './helpers/row.js'
-export type { AsyncRow, Cells, PartialRow, ResolvableRow, Row } from './helpers/row.js'
-export { rowCache } from './helpers/rowCache.js'
-export type { Selection } from './helpers/selection.js'
-export type { OrderBy } from './helpers/sort.js'
-export { resolvablePromise, wrapPromise } from './utils/promise.js'
-export type { ResolvablePromise } from './utils/promise.js'
-export { HighTable }
+import { DataFrame } from '../../helpers/dataframe.js'
+import { PartialRow } from '../../helpers/row.js'
+import { Selection, SortIndex, areAllSelected, computeNewSelection, isSelected, toggleAll } from '../../helpers/selection.js'
+import { OrderBy, areEqualOrderBy } from '../../helpers/sort.js'
+import { useInputState } from '../../hooks/useInputState.js'
+import { stringify as stringifyDefault } from '../../utils/stringify.js'
+import { throttle } from '../../utils/throttle.js'
+import Cell from '../Cell/Cell.js'
+import Row from '../Row/Row.js'
+import RowHeader from '../RowHeader/RowHeader.js'
+import TableCorner from '../TableCorner/TableCorner.js'
+import TableHeader, { cellStyle } from '../TableHeader/TableHeader.js'
+import { formatRowNumber, rowError } from './HighTable.helpers.js'
 
 /**
  * A slice of the (optionally sorted) rows to render as HTML.
