+++ src/index.ts
@@ -1,0 +1,12 @@
+import HighTable from './components/HighTable/HighTable.js'
+export { arrayDataFrame, sortableDataFrame } from './helpers/dataframe.js'
+export type { DataFrame } from './helpers/dataframe.js'
+export { asyncRows, awaitRow, awaitRows, resolvableRow } from './helpers/row.js'
+export type { AsyncRow, Cells, PartialRow, ResolvableRow, Row } from './helpers/row.js'
+export { rowCache } from './helpers/rowCache.js'
+export type { Selection } from './helpers/selection.js'
+export type { OrderBy } from './helpers/sort.js'
+export { resolvablePromise, wrapPromise } from './utils/promise.js'
+export type { ResolvablePromise } from './utils/promise.js'
+export { HighTable }
+export default HighTable
